deleted log statements, created separate interfaces module
