+++ ts/app.js
@@ -1,5 +1,6 @@
 "use strict";
 exports.__esModule = true;
+// ====== IMPORT CLASSES & INTERFACES ====== //
 var contentClasses_1 = require("./contentClasses");
 var sheetId = '11ABDt_dPctf9vJJI9LXObufyE9YsFU5nBC0Q-ul1SDs';
 var projectsAsJSON = "https://spreadsheets.google.com/feeds/list/" + sheetId + "/1/public/values?alt=json";
@@ -42,7 +43,42 @@ $(document).ready(function () {
         });
     });
 });
-//////// RENDER PAGE ELEMENTS ////////
+/*==============
+DOM MANIPULATION
+================*/
+var $dropdownMenu = $('header ul#dropdownMenu');
+var $hamburgerButton = $('i.fas.fa-bars');
+$hamburgerButton.on('click', function () {
+    $dropdownMenu.slideToggle(500);
+});
+// Found this function here: bootstrap-menu.com/detail-smart-hide.html
+// it works by checking to see if the window's current height is < the window's last height
+//// if current height < last height, user scrolled up --> show navbar
+//// if current height > last height, user scrolled down --> hide navbar
+// detect scroll top or down
+var $navbar = $('.smart-scroll');
+// detect scroll top or down
+if ($navbar.length > 0) {
+    // check if element exists
+    var last_scroll_top_1 = 0;
+    $(window).on('scroll', function () {
+        var scroll_top = $(this).scrollTop();
+        // if the current height is less than the last height, the user scrolled up and the class scrolled-up should be added
+        if (scroll_top < last_scroll_top_1) {
+            $navbar.removeClass('scrolled-down').addClass('scrolled-up');
+            // if the current height is greater than the last height, the user scrolled down and the class scrolled-up should be added
+        }
+        else {
+            $navbar.removeClass('scrolled-up').addClass('scrolled-down');
+        }
+        last_scroll_top_1 = scroll_top;
+    });
+}
+/// SUBMIT CONTACT FORM
+/*==================================================================================================
+FUNCTIONS TO FETCH DATA FROM GOOGLE SHEETS AND RENDER NEW PAGE ELEMENTS BASED ON THE DATA RETRIEVED
+==================================================================================================*/
+// RENDER PAGE ELEMENTS
 function renderData(data) {
     if (data[0].type === 'project') {
         data.forEach(function (row, index) {
@@ -71,6 +107,7 @@ function renderData(data) {
         });
     }
 }
+// make an AJAX call to the google sheets API and return a blog or project object
 function getDataFromSheet(sheet) {
     return $.ajax({ url: sheet }).then(function (data) {
         var rows;
@@ -101,35 +138,3 @@ function getDataFromSheet(sheet) {
         return rows;
     });
 }
-/////////////////////////////////////////////////////////////
-/////////////////// DOM MANIPULATION ////////////////////////
-/////////////////////////////////////////////////////////////
-var $dropdownMenu = $('header ul#dropdownMenu');
-var $hamburgerButton = $('i.fas.fa-bars');
-$hamburgerButton.on('click', function () {
-    $dropdownMenu.slideToggle(500);
-});
-//Found this function here: bootstrap-menu.com/detail-smart-hide.html
-// the way it works is by checking for the navbar's height
-// // add padding top to show content behind navbar
-// https: $('body').css('padding-top', $('.navbar').outerHeight() + 'px');
-var $navbar = $('.smart-scroll');
-// detect scroll top or down
-if ($navbar.length > 0) {
-    // check if element exists
-    var last_scroll_top_1 = 0;
-    $(window).on('scroll', function () {
-        var scroll_top = $(this).scrollTop();
-        if (scroll_top < last_scroll_top_1) {
-            $navbar.removeClass('scrolled-down').addClass('scrolled-up');
-        }
-        else {
-            $navbar.removeClass('scrolled-up').addClass('scrolled-down');
-        }
-        last_scroll_top_1 = scroll_top;
-    });
-}
-/// SUBMIT CONTACT FORM
-//////////////////////////////////////////////////////////////////////////////////////////////////////////
-// FUNCTIONS TO FETCH DATA FROM GOOGLE SHEETS AND RENDER NEW PAGE ELEMENTS BASED ON THE DATA RETRIEVED //
-////////////////////////////////////////////////////////////////////////////////////////////////////////
