+++ ts/interfaces.js
@@ -1,0 +1,2 @@
+"use strict";
+exports.__esModule = true;
